backend/requirements.txt
alembic==1.14.1
blinker==1.9.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
cmdstanpy==1.2.5
colorama==0.4.6
contourpy==1.3.1
cycler==0.12.1
filelock==3.17.0
Flask==3.1.0
Flask-Cors==5.0.0
Flask-Migrate==4.1.0
Flask-MySQLdb==2.0.0
Flask-SQLAlchemy==3.1.1
fonttools==4.56.0
fsspec==2025.2.0
greenlet==3.1.1
holidays==0.66
huggingface-hub==0.28.1
idna==3.10
importlib_resources==6.5.2
itsdangerous==2.2.0
Jinja2==3.1.5
joblib==1.4.2
kiwisolver==1.4.8
Mako==1.3.9
MarkupSafe==3.0.2
matplotlib==3.10.0
mpmath==1.3.0
mysqlclient==2.2.7
networkx==3.4.2
numpy==2.2.3
packaging==24.2
panda==0.3.1
pandas==2.2.3
patsy==1.0.1
pillow==11.1.0
prophet==1.1.6
PyMySQL==1.1.1
pyparsing==3.2.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.1
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
safetensors==0.5.2
scikit-learn==1.6.1
scipy==1.15.1
setuptools==75.8.0
six==1.17.0
SQLAlchemy==2.0.38
stanio==0.5.1
statsmodels==0.14.4
sympy==1.13.1
threadpoolctl==3.5.0
tokenizers==0.21.0
torch==2.6.0
tqdm==4.67.1
transformers==4.48.3
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
Werkzeug==3.1.3